+++ R/format_data.R
@@ -38,8 +38,18 @@
 #'   `[colname_1] > 100 & [colname_2] < 50`).
 #' @param decimals An option to specify the exact number of decimal places to
 #'   use. The default number of decimal places is `2`.
+#' @param n_sigfig A option to format numbers to *n* significant figures. By
+#'   default, this is `NULL` and thus number values will be formatted according
+#'   to the number of decimal places set via `decimals`. If opting to format
+#'   according to the rules of significant figures, `n_sigfig` must be a number
+#'   greater than or equal to `1`. Any values passed to the `decimals` and
+#'   `drop_trailing_zeros` arguments will be ignored.
 #' @param drop_trailing_zeros A logical value that allows for removal of
 #'   trailing zeros (those redundant zeros after the decimal mark).
+#' @param drop_trailing_dec_mark A logical value that determines whether decimal
+#'   marks should always appear even if there are no decimal digits to display
+#'   after formatting (e.g, `23` becomes `23.`). The default for this is `TRUE`,
+#'   which means that trailing decimal marks are not shown.
 #' @param use_seps An option to use digit group separators. The type of digit
 #'   group separator is set by `sep_mark` and overridden if a locale ID is
 #'   provided to `locale`. This setting is `TRUE` by default.
@@ -132,7 +142,9 @@ fmt_number <- function(data,
                        columns,
                        rows = NULL,
                        decimals = 2,
+                       n_sigfig = NULL,
                        drop_trailing_zeros = FALSE,
+                       drop_trailing_dec_mark = TRUE,
                        use_seps = TRUE,
                        scale_by = 1.0,
                        suffixing = FALSE,
@@ -166,6 +178,18 @@ fmt_number <- function(data,
          call. = FALSE)
   }
 
+  # Set the `formatC_format` option according to whether number
+  # formatting with significant figures is to be performed
+  if (!is.null(n_sigfig)) {
+
+    # Stop function if `n_sigfig` does not have a valid value
+    validate_n_sigfig(n_sigfig)
+
+    formatC_format <- "fg"
+  } else {
+    formatC_format <- "f"
+  }
+
   # Pass `data`, `columns`, `rows`, and the formatting
   # functions as a function list to `fmt()`
   fmt(
@@ -184,8 +208,11 @@ fmt_number <- function(data,
           scale_x_values(suffix_df$scale_by) %>%
           # Format numeric values to character-based numbers
           format_num_to_str(
-            context = context, decimals = decimals, sep_mark = sep_mark,
-            dec_mark = dec_mark, drop_trailing_zeros = drop_trailing_zeros
+            context = context, decimals = decimals, n_sigfig = n_sigfig,
+            sep_mark = sep_mark, dec_mark = dec_mark,
+            drop_trailing_zeros = drop_trailing_zeros,
+            drop_trailing_dec_mark = drop_trailing_dec_mark,
+            format = formatC_format
           ) %>%
           # With large-number suffixing support, we paste the
           # vector of suffixes to the right of the values
@@ -317,11 +344,13 @@ fmt_scientific <- function(data,
 
         x_str <-
           x %>%
-          # Format numeric values to character-based numbers
           format_num_to_str(
-            context = context, decimals = decimals, sep_mark = sep_mark,
-            dec_mark = dec_mark, drop_trailing_zeros = drop_trailing_zeros,
-            format = "e", replace_minus_mark = FALSE
+            context = context, decimals = decimals, n_sigfig = NULL,
+            sep_mark = sep_mark, dec_mark = dec_mark,
+            drop_trailing_zeros = drop_trailing_zeros,
+            drop_trailing_dec_mark = FALSE,
+            format = "e",
+            replace_minus_mark = FALSE
           )
 
         # # Determine which values don't require the (x 10^n)
@@ -365,6 +394,7 @@ fmt_symbol <- function(data,
                        accounting = FALSE,
                        decimals = NULL,
                        drop_trailing_zeros = FALSE,
+                       drop_trailing_dec_mark = TRUE,
                        use_seps = TRUE,
                        scale_by = 1.0,
                        suffixing = FALSE,
@@ -419,7 +449,8 @@ fmt_symbol <- function(data,
             # Format numeric values to character-based numbers
             format_num_to_str_c(
               context = context, decimals = decimals, sep_mark = sep_mark,
-              dec_mark = dec_mark, drop_trailing_zeros = drop_trailing_zeros
+              dec_mark = dec_mark, drop_trailing_zeros = drop_trailing_zeros,
+              drop_trailing_dec_mark = drop_trailing_dec_mark
             )
         }
 
@@ -433,7 +464,8 @@ fmt_symbol <- function(data,
             # Format numeric values to character-based numbers
             format_num_to_str_c(
               context = context, decimals = decimals, sep_mark = sep_mark,
-              dec_mark = dec_mark, drop_trailing_zeros = drop_trailing_zeros
+              dec_mark = dec_mark, drop_trailing_zeros = drop_trailing_zeros,
+              drop_trailing_dec_mark = drop_trailing_dec_mark
             )
         }
 
@@ -529,6 +561,7 @@ fmt_percent <- function(data,
                         rows = NULL,
                         decimals = 2,
                         drop_trailing_zeros = FALSE,
+                        drop_trailing_dec_mark = TRUE,
                         scale_values = TRUE,
                         use_seps = TRUE,
                         pattern = "{x}",
@@ -566,6 +599,7 @@ fmt_percent <- function(data,
     accounting = FALSE,
     decimals = decimals,
     drop_trailing_zeros = drop_trailing_zeros,
+    drop_trailing_dec_mark = drop_trailing_dec_mark,
     use_seps = use_seps,
     scale_by = scale_by,
     suffixing = FALSE,
@@ -696,6 +730,7 @@ fmt_currency <- function(data,
                          use_subunits = TRUE,
                          accounting = FALSE,
                          decimals = NULL,
+                         drop_trailing_dec_mark = TRUE,
                          use_seps = TRUE,
                          scale_by = 1.0,
                          suffixing = FALSE,
@@ -739,6 +774,7 @@ fmt_currency <- function(data,
     accounting = accounting,
     decimals = decimals,
     drop_trailing_zeros = FALSE,
+    drop_trailing_dec_mark = drop_trailing_dec_mark,
     use_seps = use_seps,
     scale_by = scale_by,
     suffixing = suffixing,